refactor: wip modal and form changes
