+++ dev-client/src/App.tsx
@@ -47,9 +47,9 @@ LogBox.ignoreLogs([
   'In React 18, SSRProvider is not necessary and is a noop. You can remove it from your app.',
 ]);
 
-function App(): JSX.Element {
-  const store = useMemo(createStore, []);
+const store = createStore();
 
+function App(): JSX.Element {
   useEffect(() =>
     checkAndroidPermissions(
       PermissionsAndroid.PERMISSIONS.ACCESS_FINE_LOCATION,
