+++ dev-client/src/components/common/modals/BottomSheetModal.tsx
@@ -1,3 +1,20 @@
+/*
+ * Copyright © 2023 Technology Matters
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see https://www.gnu.org/licenses/.
+ */
+
 import {forwardRef, useImperativeHandle, useMemo, useRef} from 'react';
 import {
   BottomSheetScrollView,
@@ -6,7 +23,12 @@ import {
 import {useHeaderHeight} from 'terraso-mobile-client/context/HeaderHeightContext';
 import {CardCloseButton} from 'terraso-mobile-client/components/common/Card';
 import {Pressable} from 'react-native';
-import {ModalMethods, Props, BackdropComponent, ModalContext} from './Modal';
+import {
+  ModalMethods,
+  Props,
+  BackdropComponent,
+  ModalContext,
+} from 'terraso-mobile-client/components/common/modals/Modal';
 
 export const BottomSheetModal = forwardRef<ModalMethods, Props>(
   ({children, trigger}, forwardedRef) => {
