+++ dev-client/src/components/dataInputs/EditIntervalModalContent.tsx
@@ -27,10 +27,10 @@ import {
 import {fromEntries} from 'terraso-client-shared/utils';
 import {useMemo, useCallback} from 'react';
 import {
-  intervalSchema,
   IntervalForm,
   IntervalFormInput,
 } from 'terraso-mobile-client/components/dataInputs/IntervalForm';
+import {intervalSchema} from 'terraso-mobile-client/components/dataInputs/intervalSchema';
 import * as yup from 'yup';
 import {useTranslation} from 'react-i18next';
 import {Heading, Row, Box, Button} from 'native-base';
@@ -39,7 +39,7 @@ import {
   FormCheckbox,
   FormSwitch,
 } from 'terraso-mobile-client/components/common/Form';
-import {useModal} from 'terraso-mobile-client/components/modals/Modal';
+import {useModal} from 'terraso-mobile-client/components/common/modals/Modal';
 
 type EditIntervalFormInput = IntervalFormInput &
   Omit<SoilDataDepthInterval, 'label' | 'depthInterval'> & {
@@ -51,7 +51,7 @@ type Props = {
   depthInterval: DepthInterval;
 };
 
-export const EditIntervalModal = ({siteId, depthInterval}: Props) => {
+export const EditIntervalModalContent = ({siteId, depthInterval}: Props) => {
   const {t} = useTranslation();
   const soilData = useSelector(state => state.soilId.soilData[siteId]);
   const dispatch = useDispatch();
