+++ dev-client/src/components/dataInputs/IntervalForm.tsx
@@ -15,57 +15,10 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-import * as yup from 'yup';
 import {FORM_LABEL_MAX} from 'terraso-mobile-client/constants';
-import {DepthInterval} from 'terraso-client-shared/soilId/soilIdSlice';
-import {TFunction} from 'i18next';
 import {useTranslation} from 'react-i18next';
 import {FormInput} from 'terraso-mobile-client/components/common/Form';
 import {Box, Row} from 'native-base';
-
-type Args = {
-  t: TFunction;
-  existingIntervals: DepthInterval[];
-};
-export const intervalSchema = ({t, existingIntervals}: Args) =>
-  yup.object({
-    label: yup
-      .string()
-      .max(
-        FORM_LABEL_MAX,
-        t('soil.depth_interval.label_help', {max: FORM_LABEL_MAX}),
-      ),
-    start: yup
-      .number()
-      .min(0)
-      .required()
-      .test((start, {createError}) => {
-        if (
-          existingIntervals.some(
-            interval => start >= interval.start && start < interval.end,
-          )
-        ) {
-          return createError({message: 'overlap'});
-        }
-        return true;
-      }),
-    end: yup
-      .number()
-      .max(200)
-      .required()
-      .test((end, {createError, parent}) => {
-        if (
-          existingIntervals.some(
-            interval => end > interval.start && end <= interval.end,
-          )
-        ) {
-          return createError({message: 'overlap'});
-        } else if (parent.start >= end) {
-          return createError({message: 'ordering'});
-        }
-        return true;
-      }),
-  });
 
 export type IntervalFormInput = {
   label: string;
