+++ dev-client/src/components/projects/ProjectInputTab.tsx
@@ -33,7 +33,7 @@ import {
   updateProjectSoilSettings,
 } from 'terraso-client-shared/soilId/soilIdSlice';
 import {Icon, IconButton} from 'terraso-mobile-client/components/common/Icons';
-import {Modal} from 'terraso-mobile-client/components/modals/Modal';
+import {Modal} from 'terraso-mobile-client/components/common/modals/Modal';
 import {AddIntervalModal} from 'terraso-mobile-client/components/dataInputs/AddIntervalModal';
 import {useMemo, useCallback} from 'react';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
