+++ dev-client/src/screens/AddUserToProjectScreen/components/MembershipControlList.tsx
@@ -15,22 +15,10 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-import {
-  Divider,
-  FlatList,
-  HStack,
-  Image,
-  Select,
-  Text,
-  VStack,
-} from 'native-base';
-import {User} from 'terraso-client-shared/account/accountSlice';
-import {Icon, IconButton} from 'terraso-mobile-client/components/common/Icons';
-import {formatName} from 'terraso-mobile-client/util';
-import {UserRole} from 'terraso-client-shared/graphqlSchema/graphql';
-import {useTranslation} from 'react-i18next';
+import {Divider, FlatList} from 'native-base';
 
-type UserFields = Omit<User, 'preferences'>;
+import {UserRole} from 'terraso-client-shared/graphqlSchema/graphql';
+import {UserDisplay} from 'terraso-mobile-client/screens/AddUserToProjectScreen/components/UserDisplay';
 
 export type UserWithRole = {user: UserFields; role: UserRole};
 
@@ -38,58 +26,6 @@ type Props = {
   users: UserWithRole[];
   updateUserRole: (role: UserRole, userId: string) => void;
   removeUser: (userId: string) => void;
-};
-
-type DisplayProps = {
-  user: UserFields;
-  role: UserRole;
-  roles: [UserRole, string][];
-  updateUserRole: (role: UserRole) => void;
-  removeUser: () => void;
-};
-
-const UserDisplay = ({
-  user: {profileImage, firstName, lastName, email},
-  roles,
-  role,
-  updateUserRole,
-  removeUser,
-}: DisplayProps) => {
-  const {t} = useTranslation();
-  return (
-    <VStack space="5px">
-      <HStack mt="15px">
-        <Image
-          variant="profilePic"
-          source={{uri: profileImage}}
-          alt={t('general.profile_image_alt')}
-        />
-        <VStack flexGrow={2} ml="15px">
-          <Text fontWeight={700} fontSize="16px">
-            {formatName(firstName, lastName)}
-          </Text>
-          <Text fontWeight={400} fontSize="14px">
-            {email}
-          </Text>
-        </VStack>
-        <IconButton name="delete" onPress={removeUser} />
-      </HStack>
-      <Select
-        width="60%"
-        ml="50px"
-        variant="unstyled"
-        dropdownCloseIcon={<Icon name="arrow-drop-down" />}
-        dropdownOpenIcon={<Icon name="arrow-drop-down" />}
-        selectedValue={role}
-        onValueChange={value => {
-          updateUserRole(value as UserRole);
-        }}>
-        {roles.map(([roleName, label]) => (
-          <Select.Item value={roleName} label={label} key={roleName} />
-        ))}
-      </Select>
-    </VStack>
-  );
 };
 
 export const MembershipControlList = ({
